requirements.txt
android-backup==0.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
construct==2.10.68
croniter==1.3.7
cryptography==38.0.1
defusedxml==0.7.1
idna==3.4
ifaddr==0.2.0
micloud==0.5
netifaces==0.11.0
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.13.0
python-dateutil==2.8.2
python-miio==0.5.12
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rich==12.6.0
shellingham==1.5.0
six==1.16.0
tqdm==4.64.1
typer==0.6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
zeroconf==0.39.1
